requirements.txt
fastapi
hypercorn
sentence-transformers
numpy
uvicorn
python-dotenv